apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
passlib[argon2]==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
qdrant-client==1.13.3
httpx==0.28.1
email-validator==2.2.0
pypdf==5.4.0
python-docx==1.1.2
pillow==11.1.0
pytesseract==0.3.13
alembic==1.15.1
